Update README.md
You will create an implementation of a linked list, one of the fundamental data structures in computer science! Under the hood, your operating system, and countless everyday applications rely upon these.

Solution Description

We highly recommend that you read this entire document in its entirety and familiarize yourself with the requirements of this assignment before writing any code!

Important: All classes created in this assignment should be generic classes. All instances of generic classes should specify a concrete type and should never use raw types or Object. If a method returns a reference type, it must use generic types, and should never use raw types or Object.

A linked list is an ordered sequence of nodes. Each node stores data and a reference to the next node in the sequence. The first node is called the head and a reference to it should be kept. This is the “entry point” to the linked list. When there are no nodes in the list, the head should be null.

Important Notes (Don’t Skip)

Non-compiling files will receive a 0 for all associated rubric items.

Do not submit .class files.